Add infrastructure for testing custom facts
This adds testing of custom facts by placing them in spec/facts, similar
to how classes have spec/classes and defines have spec/defines. The
subject is also set in the same way. It is possible to stub other facts
using a facts block and facts are properly cleared before and after a
run.