(MAINT) Disable new Rspec/RemoveConst cop
Following the recent introduction of a new cop in Rubocop,
Rspec/RemoveConst, our CI has started failing. It seems like this rule
is being enforced as a good practice to avoid buggy debug information
but no real alternative is being offered.

This commit disables the cop to allow the CI to continue working. Here
are the discussions that led to the creation of the rule:
https://github.com/rubocop/rubocop-rspec/issues/1748
https://github.com/rubocop/rubocop-rspec/pull/1749